still need to finish rework of action/routines
