+++ diagnostic/sample_temp.py
