+++ src/launch/__init__.py
@@ -10,5 +10,6 @@ Copyright (c):  2018 Danyal Ahsanullah
 License: N/A
 Description: 
 """
+
 from .cli_parser import parser
 from .command_map import cmds
