--- src/libs/states/__init__.py
@@ -1,12 +1,0 @@
-#! /usr/bin/env python
-# vim:fileencoding=utf-8
-# -*- coding: utf-8 -*-
-"""
-pi_control
-__init__.py.py
-Author: Danyal Ahsanullah
-Date: 6/27/2018
-Copyright (c):  2018 Danyal Ahsanullah
-License: N/A
-Description: 
-"""
